# src/redoc/WebApi.Service/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Sample Web API service",
    "description": "---\r\nThis Web API provides a minimal, modern ASP.NET Core service for managing a collection of games. It demonstrates clean endpoint design, lightweight configuration, and a straightforward CRUD workflow suitable for small services or as a foundation for larger systems.\r\n\r\nTo run the service locally or in a container:\r\n\r\n- Clone the repository to your development machine.\r\n- Open the solution using Visual Studio, Rider, or your preferred IDE.\r\n- Run the project directly, or build and deploy it using Docker to start the API service.\r\n\r\nToken issuer deatils:\r\n- Authority: `https://localhost:5003/connect`\r\n- OIDC configuration: `https://localhost:5003/connect/.well-known/openid-configuration`\r\n",
    "termsOfService": "https://github.com/eric-git/WebApi?tab=security-ov-file",
    "contact": {
      "name": "Eric Wu",
      "url": "https://github.com/eric-git/WebApi",
      "email": "wu_yuqing@hotmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/eric-git/WebApi?tab=MIT-1-ov-file"
    },
    "version": "v1",
    "x-logo": {
      "url": "/docs/logo.png",
      "altText": "Project logo"
    },
    "summary": "Game Web API"
  },
  "servers": [
    {
      "url": "https://localhost:5004/"
    }
  ],
  "paths": {
    "/games": {
      "get": {
        "tags": [
          "Games"
        ],
        "summary": "List all games",
        "description": "Returns all games stored in the system.",
        "operationId": "ListGames",
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://webapisuite.projecteric.io/error/500",
                  "title": "Internal Server Error",
                  "status": 500,
                  "detail": "An unexpected error occurred.",
                  "instance": "games/",
                  "traceId": "0HNIV5A3KTS4L:00000005"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://webapisuite.projecteric.io/error/401",
                  "title": "Unauthorized",
                  "status": 401,
                  "detail": "Authentication is required to access this resource.",
                  "instance": "games/",
                  "traceId": "0HNIV5A3KTS4L:00000001"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://webapisuite.projecteric.io/error/403",
                  "title": "Forbidden",
                  "status": 403,
                  "detail": "You do not have permission to access this resource.",
                  "instance": "games/",
                  "traceId": "0HNIV5A3KTS4L:00000003"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GameListItem"
                  }
                },
                "example": [
                  {
                    "Id": "76269915-5f87-4ab0-9d8a-4312b6157967",
                    "Type": "game",
                    "Name": "Mario Kart 8 Deluxe",
                    "PlayerName": "Mario",
                    "PlayerHealth": 100
                  }
                ]
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "post": {
        "tags": [
          "Games"
        ],
        "summary": "Create a new game",
        "description": "Creates a new game and returns its identifier.",
        "operationId": "CreateGame",
        "requestBody": {
          "description": "The data of the game to be created",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGame"
              },
              "example": {
                "Type": "game",
                "Name": "Mario Kart 8 Deluxe",
                "PlayerName": "Mario",
                "PlayerHealth": 100
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://webapisuite.projecteric.io/error/500",
                  "title": "Internal Server Error",
                  "status": 500,
                  "detail": "An unexpected error occurred.",
                  "instance": "games/",
                  "traceId": "0HNIV5A3KTS4L:00000005"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://webapisuite.projecteric.io/error/401",
                  "title": "Unauthorized",
                  "status": 401,
                  "detail": "Authentication is required to access this resource.",
                  "instance": "games/",
                  "traceId": "0HNIV5A3KTS4L:00000001"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://webapisuite.projecteric.io/error/403",
                  "title": "Forbidden",
                  "status": 403,
                  "detail": "You do not have permission to access this resource.",
                  "instance": "games/",
                  "traceId": "0HNIV5A3KTS4L:00000003"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "headers": {
              "Location": {
                "description": "The URL of the newly created game.",
                "schema": {
                  "type": "string",
                  "format": "uri"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                },
                "example": {
                  "type": "https://webapisuite.projecteric.io/error/400",
                  "title": "One or more validation errors occurred.",
                  "status": 400,
                  "instance": "games/",
                  "errors": {
                    "fieldName": [
                      "The field is required."
                    ]
                  },
                  "traceId": "0HNIV5A3KTS4L:00000002"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/games/{gameId}/relations": {
      "get": {
        "tags": [
          "Relations"
        ],
        "summary": "List relations for a game",
        "description": "Returns all relations associated with the specified game.",
        "operationId": "ListRelations",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "description": "The ID of the game",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://webapisuite.projecteric.io/error/500",
                  "title": "Internal Server Error",
                  "status": 500,
                  "detail": "An unexpected error occurred.",
                  "instance": "games/{gameId:guid}/relations/",
                  "traceId": "0HNIV5A3KTS4L:00000005"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://webapisuite.projecteric.io/error/401",
                  "title": "Unauthorized",
                  "status": 401,
                  "detail": "Authentication is required to access this resource.",
                  "instance": "games/{gameId:guid}/relations/",
                  "traceId": "0HNIV5A3KTS4L:00000001"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://webapisuite.projecteric.io/error/403",
                  "title": "Forbidden",
                  "status": 403,
                  "detail": "You do not have permission to access this resource.",
                  "instance": "games/{gameId:guid}/relations/",
                  "traceId": "0HNIV5A3KTS4L:00000003"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://webapisuite.projecteric.io/error/404",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "The requested resource was not found.",
                  "instance": "games/{gameId:guid}/relations/",
                  "traceId": "0HNIV5A3KTS4L:00000004"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RelationListItem"
                  }
                },
                "example": [
                  {
                    "Id": "89d39c40-5f18-43a5-8f2b-47154c8e36a2",
                    "Type": "Quest",
                    "Name": "Win Rainbow Road"
                  }
                ]
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "post": {
        "tags": [
          "Relations"
        ],
        "summary": "Create a relation",
        "description": "Creates a new relation under the specified game.",
        "operationId": "CreateRelation",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "description": "The ID of the game",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "The data of the relation to be created",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRelation"
              },
              "example": {
                "Type": "Quest",
                "Name": "Win Rainbow Road",
                "Attributes": {
                  "status": "In Progress",
                  "reward": "Gold Trophy"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://webapisuite.projecteric.io/error/500",
                  "title": "Internal Server Error",
                  "status": 500,
                  "detail": "An unexpected error occurred.",
                  "instance": "games/{gameId:guid}/relations/",
                  "traceId": "0HNIV5A3KTS4L:00000005"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://webapisuite.projecteric.io/error/401",
                  "title": "Unauthorized",
                  "status": 401,
                  "detail": "Authentication is required to access this resource.",
                  "instance": "games/{gameId:guid}/relations/",
                  "traceId": "0HNIV5A3KTS4L:00000001"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://webapisuite.projecteric.io/error/403",
                  "title": "Forbidden",
                  "status": 403,
                  "detail": "You do not have permission to access this resource.",
                  "instance": "games/{gameId:guid}/relations/",
                  "traceId": "0HNIV5A3KTS4L:00000003"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://webapisuite.projecteric.io/error/404",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "The requested resource was not found.",
                  "instance": "games/{gameId:guid}/relations/",
                  "traceId": "0HNIV5A3KTS4L:00000004"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "headers": {
              "Location": {
                "description": "The URL of the newly created relation.",
                "schema": {
                  "type": "string",
                  "format": "uri"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                },
                "example": {
                  "type": "https://webapisuite.projecteric.io/error/400",
                  "title": "One or more validation errors occurred.",
                  "status": 400,
                  "instance": "games/{gameId:guid}/relations/",
                  "errors": {
                    "fieldName": [
                      "The field is required."
                    ]
                  },
                  "traceId": "0HNIV5A3KTS4L:00000002"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateGame": {
        "required": [
          "type",
          "name",
          "playerName"
        ],
        "type": "object",
        "properties": {
          "type": {
            "maxLength": 50,
            "type": "string",
            "description": "The type of the game"
          },
          "name": {
            "maxLength": 500,
            "type": "string",
            "description": "The name of the game"
          },
          "playerName": {
            "maxLength": 255,
            "type": "string",
            "description": "The name of the game player"
          },
          "playerHealth": {
            "maximum": 100,
            "minimum": 0,
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "The health of the game player",
            "format": "int32"
          }
        }
      },
...
      "GameListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The ID of the game",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "description": "The type of the game"
          },
          "name": {
            "type": "string",
            "description": "The name of the game"
          },
          "playerName": {
            "type": "string",
            "description": "The name of the game player"
          },
          "playerHealth": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "The health of the game player",
            "format": "int32"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "RelationListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The ID of the relation",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "description": "The type of the relation"
          },
          "name": {
            "type": "string",
            "description": "The name of the relation"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT authorization header using the bearer scheme.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}